Package reorganization
- collect most storage-specific code in dss/storage/
- collect Elasticsearch-related code in dss/index/backends/es/
- promote FQID module (storage/bundles.py) to dss/
