+++ daemons/dss-checkout-sfn/app.py
@@ -13,8 +13,8 @@ from dss.logging import configure_daemon_logging
 
 from dss.stepfunctions.checkout.checkout_states import state_machine_def
 from dss.util.email import send_checkout_success_email, send_checkout_failure_email
-from dss.util.checkout import (parallel_copy, get_dst_bundle_prefix, get_manifest_files,
-                               validate_file_dst, pre_exec_validate)
+from dss.storage.checkout import (parallel_copy, get_dst_bundle_prefix, get_manifest_files,
+                                  validate_file_dst, pre_exec_validate)
 
 
 logger = logging.getLogger(__name__)
