+++ dss/events/handlers/index.py
@@ -6,9 +6,9 @@ from urllib.parse import unquote
 from abc import ABCMeta, abstractmethod
 
 from dss import Replica, Config
-from dss.storage.bundles import ObjectIdentifier, BundleFQID, TombstoneID, ObjectIdentifierError, FileFQID
-from dss.storage.index_document import BundleDocument, BundleTombstoneDocument
-from dss.util.es import elasticsearch_retry
+from dss.identifiers import ObjectIdentifier, BundleFQID, TombstoneID, ObjectIdentifierError, FileFQID
+from dss.index.backends.es.index_document import BundleDocument, BundleTombstoneDocument
+from dss.index.backends.es.__init__ import elasticsearch_retry
 
 
 logger = logging.getLogger(__name__)
