+++ dss/index/backends/es/index_document.py
@@ -18,12 +18,12 @@ from requests_http_signature import HTTPSignatureAuth
 
 from dss import Config, DeploymentStage, ESIndexType, ESDocType
 from dss import Replica
-from dss.hcablobstore import BundleMetadata, BundleFileMetadata
-from dss.storage.bundles import ObjectIdentifier, BundleFQID, TombstoneID
-from dss.storage.index import IndexManager
-from dss.storage.validator import scrub_index_data
+from dss.storage.hcablobstore import BundleMetadata, BundleFileMetadata
+from dss.identifiers import ObjectIdentifier, BundleFQID, TombstoneID
+from dss.index.backends.es.index_manager import IndexManager
+from dss.index.metadata.validator import scrub_index_data
 from dss.util import create_blob_key
-from dss.util.es import ElasticsearchClient, elasticsearch_retry
+from dss.index.backends.es.__init__ import ElasticsearchClient, elasticsearch_retry
 
 logger = logging.getLogger(__name__)
