+++ dss/storage/checkout.py
@@ -6,10 +6,10 @@ from cloud_blobstore import BlobNotFoundError, BlobStoreUnknownError
 from cloud_blobstore.s3 import S3BlobStore
 from enum import Enum, auto
 
-from .bundles import get_bundle, get_bundle_from_bucket
-from .. import DSSException, stepfunctions
-from ..config import Config, Replica
-from ..stepfunctions import s3copyclient
+from dss.storage.bundles import get_bundle, get_bundle_from_bucket
+from dss import DSSException, stepfunctions
+from dss.config import Config, Replica
+from dss.stepfunctions import s3copyclient
 
 
 log = logging.getLogger(__name__)
