+++ tests/test_object_identifier.py
@@ -7,8 +7,7 @@ from tests.infra import testmode
 pkg_root = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))  # noqa
 sys.path.insert(0, pkg_root)  # noqa
 
-import dss
-from dss.storage.bundles import ObjectIdentifier, TombstoneID, FileFQID, BundleFQID, BUNDLE_PREFIX, FILE_PREFIX
+from dss.identifiers import ObjectIdentifier, TombstoneID, FileFQID, BundleFQID, BUNDLE_PREFIX, FILE_PREFIX
 
 
 @testmode.standalone
